+++ Autopilot.h
@@ -22,21 +22,21 @@ public:
   void setModeOutdoorHover();
   void setAutopilotMode(int);
   
-  void setAltitudeKp(int8_t);
-  void setAltitudeKi(int8_t);
-  void setAltitudeBase(int8_t);
+  void setAltitudeKp(float);
+  void setAltitudeKi(float);
+  void setAltitudeBase(float);
   void setAltitudeReference(float);
   void setAltitudeTolerance(float);
 
-  void setXPositionKp(int8_t);
-  void setXPositionKi(int8_t);
-  void setXPositionKd(int8_t);
+  void setXPositionKp(float);
+  void setXPositionKi(float);
+  void setXPositionKd(float);
   void setXPositionReference(float);
   void setXPositionTolerance(float);
 
-  void setYPositionKp(int8_t);
-  void setYPositionKi(int8_t);
-  void setYPositionKd(int8_t);
+  void setYPositionKp(float);
+  void setYPositionKi(float);
+  void setYPositionKd(float);
   void setYPositionReference(float);
   void setYPositionTolerance(float);
 
@@ -49,21 +49,21 @@ public:
   void sonarXPositionCallback(void (*function)(float));
   void sonarYPositionCallback(void (*function)(float));
 
-  void altitudeKpCallback(void(*function)(int16_t));
-  void altitudeKiCallback(void(*function)(int16_t));
-  void altitudeBaseCallback(void(*function)(int16_t));
+  void altitudeKpCallback(void(*function)(float));
+  void altitudeKiCallback(void(*function)(float));
+  void altitudeBaseCallback(void(*function)(float));
   void altitudeReferenceCallback(void(*function)(float));
   void altitudeToleranceCallback(void(*function)(float));
 
-  void XPositionKpCallback(void(*function)(int16_t));
-  void XPositionKiCallback(void(*function)(int16_t));
-  void XPositionKdCallback(void(*function)(int16_t));
+  void XPositionKpCallback(void(*function)(float));
+  void XPositionKiCallback(void(*function)(float));
+  void XPositionKdCallback(void(*function)(float));
   void XPositionReferenceCallback(void(*function)(float));
   void XPositionToleranceCallback(void(*function)(float));
 
-  void YPositionKpCallback(void(*function)(int16_t));
-  void YPositionKiCallback(void(*function)(int16_t));
-  void YPositionKdCallback(void(*function)(int16_t));
+  void YPositionKpCallback(void(*function)(float));
+  void YPositionKiCallback(void(*function)(float));
+  void YPositionKdCallback(void(*function)(float));
   void YPositionReferenceCallback(void(*function)(float));
   void YPositionToleranceCallback(void(*function)(float));
 
@@ -92,27 +92,27 @@ public:
 
   
   //sync getters
-  int16_t getAltitudeKpSync();
-  int16_t getAltitudeKiSync();
-  int16_t getAltitudeBaseSync();
+  float getAltitudeKpSync();
+  float getAltitudeKiSync();
+  float getAltitudeBaseSync();
   float getAltitudeReferenceSync();
   float getAltitudeToleranceSync();
 
-  int16_t getXPositionKpSync();
-  int16_t getXPositionKiSync();
-  int16_t getXPositionKdSync();
+  float getXPositionKpSync();
+  float getXPositionKiSync();
+  float getXPositionKdSync();
   float getXPositionReferenceSync();
   float getXPositionToleranceSync();
 
-  int16_t getYPositionKpSync();
-  int16_t getYPositionKiSync();
-  int16_t getYPositionKdSync();
+  float getYPositionKpSync();
+  float getYPositionKiSync();
+  float getYPositionKdSync();
   float getYPositionReferenceSync();
   float getYPositionToleranceSync();
 
-  int16_t getSonarAltitudeSync();
-  int16_t getSonarXPositionSync();
-  int16_t getSonarYPositionSync();
+  float getSonarAltitudeSync();
+  float getSonarXPositionSync();
+  float getSonarYPositionSync();
 
 };
